pyproject.toml
[project]
name = "diligenceops"
version = "0.2.0"
description = "AI-Powered Due Diligence Pipeline for M&A"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langgraph>=0.2.0",
    "pydantic>=2.0",
    "pandas>=2.2.0",
    "httpx>=0.28.0",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
    "websockets>=14.0",
    "edgartools>=5.13.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25.0",
    "pytest-mock>=3.14.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
